+++ disambiguation/src/main/java/pl/edu/icm/coansys/disambiguation/clustering/strategies/ClusteringStrategy.java
@@ -19,6 +19,6 @@ public interface ClusteringStrategy extends Cloneable{
 	 * @return An array containing cluster numbers assigned to objects. 
 	 * Two objects sharing the same cluster number may be considered as the same one.  
 	 */
-    int[] clusterize(double[][] similarities);
+    int[] clusterize(float[][] similarities);
     
 }
