+++ disambiguation/src/main/java/pl/edu/icm/coansys/disambiguation/clustering/strategies/SingleLinkageHACStrategy.java
@@ -30,7 +30,7 @@ public abstract class SingleLinkageHACStrategy implements ClusteringStrategy {
      * one.
      */
     @Override
-    public int[] clusterize(double sim[][]) {
+    public int[] clusterize(float sim[][]) {
         int[] I = new int[sim.length];
         ClusterElement[] nearBestMatch = new ClusterElement[sim.length];
 
@@ -39,7 +39,7 @@ public abstract class SingleLinkageHACStrategy implements ClusteringStrategy {
         for (int n = 0; n < sim.length; n++) {
             C[n] = new ClusterElement[n];
 
-            double maxSim = -1;
+            float maxSim = -1;
             if (C[n].length != 0) {
                 maxSim = sim[n][0];
             }
@@ -74,7 +74,7 @@ public abstract class SingleLinkageHACStrategy implements ClusteringStrategy {
             if (i1 == i2) {
                 continue;
             }
-            double simil = (i1 > i2) ? C[i1][i2].getSim() : C[i2][i1].getSim();
+            float simil = (i1 > i2) ? C[i1][i2].getSim() : C[i2][i1].getSim();
             if (simil < 0) {
                 return I;
             }
@@ -103,7 +103,7 @@ public abstract class SingleLinkageHACStrategy implements ClusteringStrategy {
     }
 
     protected int argMaxSequenceIndexExcludeSame(ClusterElement[] nearBestMatch, int[] I) {
-        double maxval = Double.NEGATIVE_INFINITY;
+        float maxval = Float.NEGATIVE_INFINITY;
         int maxvalindex = -1;
 
         for (int i = 0; i < nearBestMatch.length; i++) {
@@ -126,7 +126,7 @@ public abstract class SingleLinkageHACStrategy implements ClusteringStrategy {
 
     protected ClusterElement argMaxElementWithConstraints(ClusterElement[] Cn,
             int[] I, int forbidden) {
-        double maxval = -1;
+        float maxval = -1;
         ClusterElement retEl = null;
         for (int i = 0; i < Cn.length; i++) {
             if (i == forbidden) {
@@ -143,5 +143,5 @@ public abstract class SingleLinkageHACStrategy implements ClusteringStrategy {
         return retEl;
     }
 
-    protected abstract double SIM(double a, double b);
+    protected abstract float SIM(float a, float b);
 }
